+++ scripts/check_workspace_contract.py
@@ -69,6 +69,7 @@ EXPECTED_CRATES: tuple[str, ...] = (
     "analysis_engine",
     "topic_measurement",
     "psychometric_core",
+    "orchestrator_live",
 )
 
 REQUIRED_CI_SNIPPETS: tuple[str, ...] = (